Спринт 11. Исправления после ревью
Вынес в отдельные файлы структуры и протоколы
Использовал ленивые свойства
Внес правки по рекомендациям
