+++ Image Feed/Auth/AuthViewControllerDelegate.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol AuthViewControllerDelegate: AnyObject {
+    func authViewController(_ vc: AuthViewController, didAuthenticateWithCode code: String)
+}
