+++ Image Feed/Auth/WebViewViewControllerDelegate.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol WebViewViewControllerDelegate: AnyObject {
+    func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String)
+    func webViewViewControllerDidCancel(_ vc: WebViewViewController)
+}
