+++ Image Feed/OAuth2/OAuthTokenResponseBody.swift
@@ -1,3 +1,5 @@
+import Foundation
+
 struct OAuthTokenResponseBody: Decodable {
     let accessToken: String
     let tokenType: String
