+++ Image Feed/OAuth2/ProfileImageService.swift
@@ -1,21 +1,6 @@
 import Foundation
 import UIKit
 
-// MARK: - UserResult
-struct UserResult: Codable {
-    let profileImage: ProfileImage?
-
-    struct ProfileImage: Codable {
-        let small: String?
-        let medium: String?
-        let large: String?
-    }
-
-    enum CodingKeys: String, CodingKey {
-        case profileImage = "profile_image"
-    }
-}
-
 // MARK: - ProfileImageService
 final class ProfileImageService {
